feat: Prepend LINT to block labels
This reduces the chance of actual code being detected as a OnChange
block. The chance of code matching this is extremely low.

The alternative is to embed a syntax parser for every single supported
language to detect comments in a given code file. That is just overkill
for something like this :)